Add support for converting timestamps (#17)
All 3 targets support parsing `UNIX` and `UNIX_MS`.

ecs-mapper lets users specify any kind of shorthand or date format as well, but does not validate them. They are passed as is to the target pipelines.

This is not a compatibility layer. However this can be useful once the user already knows which kind of pipeline they want to use, and know exactly which time format they need in the resulting pipeline.